factor scope editor out into lib
